[ICU-900] Fixed some compiler warnings. X-Commit-URL: https://ssl.icu-project.org/trac/changeset/8539 Former-commit-id: 067868bcf80d6fcbdb4b8da661d340621ac19adc

+++ icu4c/source/common/locmap.c
@@ -111,7 +111,7 @@ uprv_defaultCodePageForLocale(const char *locale)
     {
         return NULL;
     }
-    locale_len = uprv_strlen(locale);
+    locale_len = (int32_t)uprv_strlen(locale);
 
     if(locale_len < 2)
     {
@@ -124,7 +124,7 @@ uprv_defaultCodePageForLocale(const char *locale)
     {
         if(uprv_strncmp(locale, _localeToDefaultCharmapTable[i].loc,
                         uprv_min(locale_len,
-                                 uprv_strlen(_localeToDefaultCharmapTable[i].loc)))
+                                 (int32_t)uprv_strlen(_localeToDefaultCharmapTable[i].loc)))
             == 0)
         {
             return _localeToDefaultCharmapTable[i].charmap;
